Almost there, time to get messy...

+++ include/factorial.h
@@ -1,15 +1,15 @@
 /**
-  * Header to src/factorial.c
-  *
-  * Created: 2013-09-26
-  * Revised: 2013-09-29
-  */
+	* Header to src/factorial.c
+	*
+	* Created: 2013-09-26
+	* Revised: 2013-09-30
+	*/
 
 #ifndef __FACTORIAL_H__
-  #define __FACTORIAL_H__
+	#define __FACTORIAL_H__
 
-//  Limited to a product no greater than 18446744073709551615
-long double
-factorial ( unsigned short start , unsigned short end ) ;
+	//  Limited to a product no greater than 18446744073709551615
+	double
+	factorial ( unsigned short start , unsigned short end ) ;
 
 #endif
